requirements.txt
streamlit
torch
torchvision
pillow
numpy
opencv-python
matplotlib
scikit-learn
pytorch-grad-cam
timm
opencv-python-headless
